pyproject.toml
[project]
name = "bidi-custom"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.22.1",  # 1.24.1で接続エラーが発生するため一時的にダウングレード
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.0",
    "websockets>=13.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "greenlet>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
